pyproject.toml
[tool.poetry]
name = "sdss-hartmannActor"
version = "3.0.0b3"
description = "An actor to analyse hartmann images and apply corrections."
authors = ["José Sánchez-Gallego <gallegoj@uw.edu>", "John Parejko <parejkoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/hartmannActor"
repository = "https://github.com/sdss/hartmannActor"
keywords = ["astronomy", "software"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "hartmann", from = "python" }
]
include = ["python/hartmann/etc/*"]

[tool.poetry.dependencies]
python = "^3.9,<3.13"
numpy = "^1.20.0"
scipy = "^1.8.0"
astropy = "^5.0"
sdsstools = ">=0.5.0"
sdss-clu = "^1.6.2"
click-default-group = "^1.2.2"
pandas = "^1.4.3"
seaborn = "^0.11.2"
matplotlib = "^3.5.2"
rich = "^12.6.0"

[tool.poetry.group.dev.dependencies]
ipython = ">=8.0.0"
invoke = ">=1.3.0"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-cov = ">=2.8.1"
pytest-sugar = ">=0.9.2"
pytest-asyncio = ">=0.19.0"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
toml = ">=0.10.0"
black = {version = ">=21.7b0", allow-prereleases = true}
ruff = ">=0.1.6"

[tool.poetry.group.docs.dependencies]
rstcheck = ">=3.3.1"
Sphinx = ">=3.0.0"
sphinx-click = ">=2.6.0"
sphinx-jsonschema = ">=1.16.7"
myst-parser = ">=0.14.0"
furo = ">=2021.6.18-beta.36"
nox = ">=2021.6.12"
sphinx-autobuild = ">=2021.3.14"
sphinx-copybutton = ">=0.3.3"

[tool.poetry.scripts]
hartmann = 'hartmann.__main__:hartmann'

[tool.black]
line-length = 88
target-version = ['py312']
fast = true

[tool.ruff]
line-length = 88
target-version = 'py312'
select = ["E", "F", "I"]
unfixable = ["F841"]
exclude = ["typings/"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403", "E402"]

[tool.ruff.isort]
known-first-party = ["hartmann"]
lines-after-imports = 2
section-order = ["future", "standard-library", "typing", "third-party", "sdss", "first-party", "local-folder"]

[tool.ruff.isort.sections]
typing = ["typing"]
sdss = ["sdsstools", "clu"]

[tool.pytest.ini_options]
addopts = "--cov hartmann --cov-report xml --cov-report html --cov-report term"
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
omit = [
    "*/__main__.py",
    "*/calibration.py"
]

[tool.coverage.report]
exclude_lines = [
    "# pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
]

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
